refactor-ir-cake: removed Library trait (merged with Scalan)
