+++ sc/shared/src/main/scala/special/sigma/wrappers/WrappersModule.scala
@@ -1,4 +1,6 @@
 package sigma.wrappers
 
+import scalan.Scalan
+
 trait WrappersModule
-  extends special.wrappers.WrappersModule
+  extends special.wrappers.WrappersModule { self: Scalan => }
