+++ sc/shared/src/main/scala/special/wrappers/WrappersModule.scala
@@ -1,5 +1,6 @@
 package special.wrappers
 
+import scalan.Scalan
 import wrappers.scala.WOptionsModule
 import wrappers.scalan.WRTypesModule
 import wrappers.special.WSpecialPredefsModule
@@ -7,4 +8,4 @@ import wrappers.special.WSpecialPredefsModule
 trait WrappersModule
   extends WSpecialPredefsModule
   with WOptionsModule
-  with WRTypesModule
+  with WRTypesModule { self: Scalan => }
