+++ sc/shared/src/main/scala/wrappers/scalan/WRTypes.scala
@@ -1,9 +1,8 @@
 package wrappers.scalan {
   import scalan._
-  import special.wrappers.WrappersModule
 
 
-  trait WRTypes extends Base { self: WrappersModule =>
+  trait WRTypes extends Base { self: Scalan =>
     trait WRType[A] extends Def[WRType[A]] {
       implicit def eA: Elem[A];
       def name: Ref[String]
