+++ sc/shared/src/test/scala/scalan/TestLibrary.scala
@@ -1,6 +1,6 @@
 package scalan
 
-trait TestLibrary extends Library {
+trait TestLibrary extends Scalan {
   import CollBuilder._
   val reflection = (GraphIRReflection)
   lazy val colBuilder: Ref[CollBuilder] = variable[CollBuilder]
